Split dao.py into serializers, queries, and domain modules
Moved all functions out of dao.py into focused modules:
- serializers.py: serialize_datetime, serialize_line_item, serialize_user, serialize_event
- queries.py: all get_* read functions
- domain.py: get_categorized_data, get_line_item_amounts, remove_event_from_line_item,
             create_manual_transaction, delete_manual_transaction

Updated all callers (resources/, application.py, tests/) to import directly
from the new modules and deleted dao.py.
