--- server/dao.py
@@ -1,478 +1,0 @@
-import logging
-from datetime import datetime
-from datetime import timezone as tz
-from decimal import Decimal
-from typing import Any, Dict, List, Optional, Union
-
-from sqlalchemy.orm import joinedload, subqueryload
-
-from models.sql_models import Event, LineItem, User
-
-logger = logging.getLogger(__name__)
-
-
-def remove_event_from_line_item(line_item_id: Union[str, int]) -> None:
-    """
-    Remove event association from a line item by its ID.
-    """
-    from models.database import SessionLocal
-    from models.sql_models import EventLineItem, LineItem
-
-    try:
-        with SessionLocal.begin() as db:
-            line_item = db.query(LineItem).filter(LineItem.id == str(line_item_id)).first()
-            if line_item:
-                db.query(EventLineItem).filter(EventLineItem.line_item_id == line_item.id).delete()
-            else:
-                logger.warning(f"Could not find line item with ID {line_item_id}")
-    except Exception as e:
-        logger.error(f"Failed to remove event from line item {line_item_id}: {e}")
-        raise
-
-
-def get_user_by_email(email: str) -> Optional[Dict[str, Any]]:
-    from models.database import SessionLocal
-    from models.sql_models import User
-
-    with SessionLocal.begin() as db:
-        user = db.query(User).filter(User.email == email).first()
-        if not user:
-            return None
-        return {
-            "id": user.id,
-            "first_name": user.first_name,
-            "last_name": user.last_name,
-            "email": user.email,
-            "password_hash": user.password_hash,
-        }
-
-
-def get_categorized_data() -> List[Dict[str, Any]]:
-    """Group totalExpense by month, year, and category.
-
-    Uses SQL subqueries to compute per-event amounts instead of loading all
-    LineItem ORM objects, reducing the work from O(events * line_items) to
-    O(events). Duplicate events (is_duplicate=True) count only the minimum
-    line item amount rather than the sum.
-    """
-    from collections import defaultdict
-    from datetime import timezone as tz
-
-    from sqlalchemy import case, func
-
-    from models.database import SessionLocal
-    from models.sql_models import Category, Event, EventLineItem, LineItem
-
-    with SessionLocal.begin() as db:
-        # Subquery: total amount per event (sum of all line items)
-        total_subq = (
-            db.query(
-                EventLineItem.event_id,
-                func.sum(LineItem.amount).label("total"),
-            )
-            .join(LineItem, EventLineItem.line_item_id == LineItem.id)
-            .group_by(EventLineItem.event_id)
-            .subquery()
-        )
-
-        # Subquery: minimum line item amount per event (for duplicates)
-        min_li_subq = (
-            db.query(
-                EventLineItem.event_id,
-                func.min(LineItem.amount).label("min_amount"),
-            )
-            .join(LineItem, EventLineItem.line_item_id == LineItem.id)
-            .group_by(EventLineItem.event_id)
-            .subquery()
-        )
-
-        # Main query: events with category and computed amount — no ORM line item loading
-        rows = (
-            db.query(
-                Event.date,
-                Category.name.label("category"),
-                case(
-                    (Event.is_duplicate == True, min_li_subq.c.min_amount),  # noqa: E712
-                    else_=total_subq.c.total,
-                ).label("amount"),
-            )
-            .join(Category, Event.category_id == Category.id)
-            .outerjoin(total_subq, total_subq.c.event_id == Event.id)
-            .outerjoin(min_li_subq, min_li_subq.c.event_id == Event.id)
-            .all()
-        )
-
-        breakdown: Dict[tuple, Decimal] = defaultdict(Decimal)
-        for row in rows:
-            if not row.category:
-                continue
-            utc_date = row.date.astimezone(tz.utc) if row.date.tzinfo else row.date.replace(tzinfo=tz.utc)
-            key = (utc_date.year, utc_date.month, row.category)
-            breakdown[key] += row.amount if row.amount is not None else Decimal("0")
-
-        return [
-            {"year": year, "month": month, "category": category, "totalExpense": float(amount)}
-            for (year, month, category), amount in sorted(breakdown.items())
-        ]
-
-
-def serialize_datetime(dt: Optional[datetime]) -> float:
-    """Treats naive datetimes from SQLite as UTC to ensure consistent timestamp conversion"""
-    if not dt:
-        return 0.0
-
-    if dt.tzinfo is None:
-        return dt.replace(tzinfo=tz.utc).timestamp()
-    else:
-        return dt.timestamp()
-
-
-def serialize_line_item(li: LineItem) -> Dict[str, Any]:
-    """Convert LineItem ORM to dict"""
-    # Determine if this is a manual transaction based on the source
-    is_manual = li.transaction.source == "manual" if li.transaction else False
-
-    data = {
-        "id": li.id,
-        "date": serialize_datetime(li.date),
-        "payment_method": li.payment_method.name if li.payment_method else "Unknown",
-        "description": li.description or "",
-        "amount": float(li.amount or 0.0),
-        "responsible_party": li.responsible_party,
-        "notes": li.notes,
-        "is_manual": is_manual,
-    }
-    if li.events:
-        data["event_id"] = li.events[0].id
-    return data
-
-
-def serialize_user(user: User) -> Dict[str, Any]:
-    """Convert User ORM to dict"""
-    return {
-        "id": user.id,
-        "first_name": user.first_name,
-        "last_name": user.last_name,
-        "email": user.email,
-        "password_hash": user.password_hash,
-    }
-
-
-def serialize_event(event: Event) -> Dict[str, Any]:
-    """Convert Event ORM to dict"""
-    amount = float(event.total_amount) if event.total_amount else 0.0
-    line_item_ids = [li.id for li in event.line_items]
-    tag_names = [tag.name for tag in event.tags] if event.tags else []
-
-    return {
-        "id": event.id,
-        "date": serialize_datetime(event.date),
-        "name": event.description or "",
-        "category": event.category.name if event.category else "Unknown",
-        "amount": amount,
-        "line_items": line_item_ids,
-        "tags": tag_names,
-        "is_duplicate_transaction": event.is_duplicate or False,
-    }
-
-
-def get_line_item_amounts(line_item_ids: list[str]) -> List[Dict[str, Any]]:
-    """Get only id, date, and amount for the given line items — no relationship loading."""
-    from models.database import SessionLocal
-    from models.sql_models import LineItem
-
-    with SessionLocal.begin() as db:
-        rows = (
-            db.query(LineItem.id, LineItem.date, LineItem.amount)
-            .filter(LineItem.id.in_([str(id) for id in line_item_ids]))
-            .all()
-        )
-        return [{"id": row.id, "date": serialize_datetime(row.date), "amount": float(row.amount or 0.0)} for row in rows]
-
-
-def get_all_line_items(
-    ids: Optional[List[str]] = None,
-    payment_method: Optional[str] = None,
-    only_unreviewed: bool = False,
-    limit: Optional[int] = None,
-    offset: int = 0,
-) -> List[Dict[str, Any]]:
-    """Get line items from PostgreSQL"""
-    from models.database import SessionLocal
-    from models.sql_models import Event, LineItem, PaymentMethod
-
-    with SessionLocal.begin() as db:
-        query = db.query(LineItem).options(
-            joinedload(LineItem.payment_method),
-            joinedload(LineItem.events),
-            joinedload(LineItem.transaction),
-        )
-
-        if ids is not None:
-            query = query.filter(LineItem.id.in_(ids))
-        if payment_method and payment_method != "All":
-            query = query.join(LineItem.payment_method).filter(PaymentMethod.name == payment_method)
-        if only_unreviewed:
-            query = query.outerjoin(LineItem.events).filter(Event.id.is_(None))
-
-        query = query.order_by(LineItem.date.desc())
-        if offset:
-            query = query.offset(offset)
-        if limit is not None:
-            query = query.limit(limit)
-        line_items = query.all()
-        return [serialize_line_item(li) for li in line_items]
-
-
-def get_line_item_by_id(id: str) -> Optional[Dict[str, Any]]:
-    """Get line item from PostgreSQL by ID"""
-    from models.database import SessionLocal
-    from models.sql_models import LineItem
-
-    with SessionLocal.begin() as db:
-        query = db.query(LineItem).options(
-            joinedload(LineItem.payment_method),
-            joinedload(LineItem.events),
-            joinedload(LineItem.transaction),
-        )
-        line_item = query.filter(LineItem.id == id).first()
-        return serialize_line_item(line_item) if line_item else None
-
-
-def get_user_by_id(id: str) -> Optional[Dict[str, Any]]:
-    """Get user from PostgreSQL by ID"""
-    from models.database import SessionLocal
-    from models.sql_models import User
-
-    with SessionLocal.begin() as db:
-        query = db.query(User)
-        user = query.filter(User.id == id).first()
-        return serialize_user(user) if user else None
-
-
-def get_all_events(
-    filters: Optional[Dict[str, Any]],
-    limit: Optional[int] = None,
-    offset: int = 0,
-) -> List[Dict[str, Any]]:
-    """Get events from PostgreSQL"""
-    from datetime import datetime
-
-    from models.database import SessionLocal
-    from models.sql_models import Event, LineItem
-
-    with SessionLocal.begin() as db:
-        # Use subqueryload for one-to-many relationships to avoid duplicate rows
-        # joinedload is fine for many-to-one (category)
-        query = db.query(Event).options(
-            joinedload(Event.category),
-            subqueryload(Event.line_items).joinedload(LineItem.payment_method),
-            subqueryload(Event.tags),
-        )
-
-        if filters and "date" in filters:
-            date_filter = filters["date"]
-            if isinstance(date_filter, dict):
-                from datetime import UTC
-
-                if "$gte" in date_filter:
-                    query = query.filter(Event.date >= datetime.fromtimestamp(date_filter["$gte"], UTC))
-                if "$lte" in date_filter:
-                    query = query.filter(Event.date <= datetime.fromtimestamp(date_filter["$lte"], UTC))
-
-        query = query.order_by(Event.date.desc())
-        if offset:
-            query = query.offset(offset)
-        if limit is not None:
-            query = query.limit(limit)
-        events = query.all()
-        return [serialize_event(event) for event in events]
-
-
-def get_event_by_id(id: str) -> Optional[Dict[str, Any]]:
-    """Get event from PostgreSQL by ID"""
-    from models.database import SessionLocal
-    from models.sql_models import Event, LineItem
-
-    with SessionLocal.begin() as db:
-        query = db.query(Event).options(
-            joinedload(Event.category),
-            subqueryload(Event.line_items).joinedload(LineItem.payment_method),
-            subqueryload(Event.tags),
-        )
-
-        event = query.filter(Event.id == id).first()
-        return serialize_event(event) if event else None
-
-
-def get_transactions(source: str, filters: Optional[Dict[str, Any]]) -> List[Dict[str, Any]]:
-    """Get raw transactions from PostgreSQL by source (venmo_api, splitwise_api, stripe_api, manual)"""
-    from models.database import SessionLocal
-    from models.sql_models import Transaction
-
-    with SessionLocal.begin() as db:
-        query = db.query(Transaction).filter(Transaction.source == source)
-
-        # Apply filters if provided (though typically raw data reads don't use filters)
-        if filters:
-            # Could add date filters here if needed in the future
-            pass
-
-        query = query.order_by(Transaction.transaction_date.desc())
-        transactions = query.all()
-
-        return [
-            {
-                "source_id": txn.source_id,
-                **txn.source_data,  # Unpack the JSONB data
-            }
-            for txn in transactions
-        ]
-
-
-def get_all_bank_accounts(
-    filters: Optional[Dict[str, Any]],
-) -> List[Dict[str, Any]]:
-    """Get bank accounts from PostgreSQL"""
-    from models.database import SessionLocal
-    from models.sql_models import BankAccount
-
-    with SessionLocal.begin() as db:
-        query = db.query(BankAccount)
-
-        if filters and "status" in filters:
-            query = query.filter(BankAccount.status == filters["status"])
-
-        accounts = query.all()
-        return [
-            {
-                "id": acc.id,
-                "institution_name": acc.institution_name,
-                "display_name": acc.display_name,
-                "last4": acc.last4,
-                "status": acc.status,
-                "can_relink": acc.can_relink,
-                "currency": acc.currency,
-                "latest_balance": float(acc.latest_balance) if acc.latest_balance is not None else None,
-                "balance_as_of": acc.balance_as_of,
-            }
-            for acc in accounts
-        ]
-
-
-def create_manual_transaction(
-    transaction_id: str,
-    line_item_id: str,
-    transaction_date: Any,
-    posix_date: float,
-    amount: Any,
-    description: str,
-    payment_method_id: str,
-    responsible_party: str,
-) -> None:
-    """
-    Create a manual transaction with its associated line item.
-
-    Manual transactions bypass the bulk upsert pipeline since they don't need
-    deduplication logic (there's no external API to re-import from).
-
-    Args:
-        transaction_id: Generated txn_xxx ID for the transaction
-        line_item_id: Generated li_xxx ID for the line item
-        transaction_date: datetime object for the transaction
-        posix_date: POSIX timestamp of the transaction date
-        amount: Decimal amount for the line item
-        description: Transaction description
-        payment_method_id: ID of the payment method
-        responsible_party: Name of the responsible party
-    """
-    from models.database import SessionLocal
-    from models.sql_models import LineItem, Transaction
-
-    try:
-        with SessionLocal.begin() as db:
-            transaction = Transaction(
-                id=transaction_id,
-                source="manual",
-                source_id=transaction_id,
-                source_data={
-                    "date": posix_date,
-                    "person": responsible_party,
-                    "description": description,
-                    "amount": float(amount),
-                    "payment_method_id": payment_method_id,
-                },
-                transaction_date=transaction_date,
-            )
-            db.add(transaction)
-
-            line_item = LineItem(
-                id=line_item_id,
-                transaction_id=transaction_id,
-                date=transaction_date,
-                amount=amount,
-                description=description,
-                payment_method_id=payment_method_id,
-                responsible_party=responsible_party,
-            )
-            db.add(line_item)
-    except Exception as e:
-        logger.error(f"Failed to create manual transaction: {e}")
-        raise
-
-
-def delete_manual_transaction(transaction_id: str) -> bool:
-    """
-    Delete a manual transaction and its associated line items.
-
-    Args:
-        transaction_id: The transaction ID to delete
-
-    Returns:
-        True if deleted, False if not found
-
-    Raises:
-        ValueError: If the transaction's line item is assigned to an event
-    """
-    from models.database import SessionLocal
-    from models.sql_models import EventLineItem, LineItem, Transaction
-
-    try:
-        with SessionLocal.begin() as db:
-            transaction = (
-                db.query(Transaction).filter(Transaction.id == transaction_id, Transaction.source == "manual").first()
-            )
-
-            if not transaction:
-                return False
-
-            line_item = db.query(LineItem).filter(LineItem.transaction_id == transaction.id).first()
-
-            if line_item:
-                is_assigned = db.query(EventLineItem).filter(EventLineItem.line_item_id == line_item.id).first()
-                if is_assigned:
-                    raise ValueError("Cannot delete transaction with line item assigned to an event")
-
-            # line item is deleted by cascade
-            db.delete(transaction)
-            return True
-    except Exception as e:
-        logger.error(f"Failed to delete manual transaction {transaction_id}: {e}")
-        raise
-
-
-def get_payment_method_by_id(payment_method_id: str) -> Optional[Dict[str, Any]]:
-    """Get payment method by ID"""
-    from models.database import SessionLocal
-    from models.sql_models import PaymentMethod
-
-    with SessionLocal.begin() as db:
-        pm = db.query(PaymentMethod).filter(PaymentMethod.id == payment_method_id).first()
-        if not pm:
-            return None
-        return {
-            "id": pm.id,
-            "name": pm.name,
-            "type": pm.type,
-            "is_active": pm.is_active,
-        }
