+++ server/resources/auth.py
@@ -12,8 +12,8 @@ from flask_jwt_extended import (
 )
 
 from constants import GATED_USERS
-from dao import get_user_by_email
 from helpers import check_password, hash_password
+from queries import get_user_by_email
 from utils.id_generator import generate_id
 from utils.pg_bulk_ops import upsert_user
