+++ server/resources/event.py
@@ -5,13 +5,9 @@ from flask import Blueprint, Response, jsonify, request
 from flask_jwt_extended import get_current_user, jwt_required
 
 from constants import LARGEST_EPOCH_TIME, SMALLEST_EPOCH_TIME
-from dao import (
-    get_all_events,
-    get_all_line_items,
-    get_event_by_id,
-    get_line_item_amounts,
-)
+from domain import get_line_item_amounts
 from helpers import html_date_to_posix
+from queries import get_all_events, get_all_line_items, get_event_by_id
 
 logger = logging.getLogger(__name__)
